requirements.txt
aiohttp>=3.9.0
gspread>=6.0.0
google-auth>=2.28.0
